Move `Dump` models to `dumps.py`
