+++ fractal_server/app/schemas/applyworkflow.py
@@ -3,11 +3,13 @@ from enum import Enum
 from typing import Optional
 
 from pydantic import BaseModel
-from pydantic import Extra
 from pydantic import validator
 from pydantic.types import StrictStr
 
 from ._validators import valstr
+from .dumps import DatasetDump
+from .dumps import ProjectDump
+from .dumps import WorkflowDump
 
 
 __all__ = (
@@ -15,55 +17,6 @@ __all__ = (
     "ApplyWorkflowCreate",
     "ApplyWorkflowRead",
 )
-
-
-class ProjectDump(BaseModel, extra=Extra.forbid):
-
-    id: int
-    name: str
-    read_only: bool
-    timestamp_created: str
-
-
-class TaskDump(BaseModel):
-    id: int
-    source: str
-    name: str
-    command: str
-    input_type: str
-    output_type: str
-    owner: Optional[str]
-    version: Optional[str]
-
-
-class WorkflowTaskDump(BaseModel):
-    id: int
-    order: Optional[int]
-    workflow_id: int
-    task_id: int
-    task: TaskDump
-
-
-class WorkflowDump(BaseModel):
-    id: int
-    name: str
-    project_id: int
-    task_list: list[WorkflowTaskDump]
-
-
-class ResourceDump(BaseModel):
-    id: int
-    path: str
-    dataset_id: int
-
-
-class DatasetDump(BaseModel):
-    id: int
-    name: str
-    type: Optional[str]
-    read_only: bool
-    resource_list: list[ResourceDump]
-    project_id: int
 
 
 class JobStatusType(str, Enum):
